Moved preferences management functionality from app delegate to a separate file.
