+++ RayRacer/RayRacerDelegate.swift
@@ -18,7 +18,7 @@ class RayRacerDelegate: NSObject, NSApplicationDelegate {
 	private var commandQueue: MTLCommandQueue!
 	private var pipelineState: MTLRenderPipelineState!
 	
-	var console = Atari2600()
+	private(set) var console = Atari2600()
 	
 	func applicationDidFinishLaunching(_ notification: Notification) {
 		guard let device = MTLCreateSystemDefaultDevice(),
@@ -52,20 +52,9 @@ class RayRacerDelegate: NSObject, NSApplicationDelegate {
 // MARK: Target actions
 extension RayRacerDelegate {
 	@IBAction func didSelectInsertCartridgeMenuItem(_ sender: AnyObject) {
-		let panel = NSOpenPanel()
-		panel.allowsMultipleSelection = false
-		panel.canChooseFiles = true
-		panel.canChooseDirectories = false
-		panel.canCreateDirectories = false
-		panel.directoryURL = self.defaults.openedFileURLs.first
-		
-		let response = panel.runModal()
-		guard let url = panel.url,
-			  response == .OK else {
-			return
-		}
-		
-		self.openFile(at: url)
+		self.withModalFileOpenPanel({
+			self.openFile(at: $0)
+		})
 	}
 	
 	@IBAction func didSelectInsertRecentCartridgeMenuItem(_ sender: NSMenuItem) {
@@ -95,27 +84,21 @@ extension RayRacerDelegate {
 	@IBAction func didSelectGameResetMenuItem(_ sender: AnyObject) {
 	}
 	
-	@IBAction func didSelectResetMenuItem(_ sender: AnyObject) {
-		self.console.reset()
-	}
-	
 	private func didSelectConsoleSwitchesMenuItem(_ menuItem: NSMenuItem, for switch: Atari2600.Switches) {
 		// menu items for `on` switch values appear at the top of the menu
 		let on = menuItem.menu?
 			.index(of: menuItem) == 0
 		
 		self.console.setSwitch(`switch`, on: on)
-		// TODO: -
-		//		self.defaults.consoleSwitches = self.console?.switches
+		self.defaults.consoleSwitches = self.console.switches
 	}
-}
-
-extension RayRacerDelegate {
+	
+	@IBAction func didSelectResetMenuItem(_ sender: AnyObject) {
+		self.console.reset()
+	}
+	
 	@IBAction func didSelectGameResumeMenuItem(_ sender: AnyObject) {
-//		if let identifier = self.console.gameIdentifier {
-//			let breakpoints = self.defaults.breakpoints(forGameIdentifier: identifier)
-//			console.resumeProgram(until: breakpoints)
-//		}
+		// TODO: resume game
 	}
 	
 	@IBAction func didSelectStepInstructionMenuItem(_ sender: AnyObject) {
@@ -150,30 +133,6 @@ extension RayRacerDelegate: NSWindowDelegate {
 		if let window = notification.object as? NSWindow {
 			self.windowControllers.remove(where: { $0.window == window })
 		}
-	}
-	
-	func openFile(at url: URL) {
-		guard let data = try? Data(contentsOfSecurityScopedResourceAt: url) else {
-			// TODO: show error when opening cartridge data fails
-			fatalError()
-		}
-		
-		let viewController = ScreenViewController(
-			commandQueue: self.commandQueue,
-			pipelineState: self.pipelineState)
-		
-		let windowController = NSWindowController(windowNibName: "ScreenWindow")
-		windowController.contentViewController = viewController
-		windowController.window?
-			.title = url.lastPathComponent
-		
-		self.console.tia.output = viewController
-		self.console.switches = self.defaults.consoleSwitches
-		self.console.insertCartridge(data)
-		self.console.reset()
-		
-		self.showWindow(of: windowController)
-		self.defaults.addOpenedFileURL(url)
 	}
 }
 
@@ -286,94 +245,47 @@ private extension NSToolbarItem.Identifier {
 }
 
 
-
 // MARK: -
-// MARK: Preference management
-private extension String {
-	static let consoleSwitches = "ConsoleSwitches"
-	static let openedFileBookmarks = "OpenedFileBookmarks"
-}
-
-private extension UserDefaults {
-	var consoleSwitches: Atari2600.Switches {
-		get {
-			// by default, TV type is set to `color` and both difficulties
-			// to `advanced`
-			let value = self.object(forKey: .consoleSwitches) as? Int ?? 0xc8
-			return Atari2600.Switches(rawValue: value)
-		}
-		set {
-			self.setValue(newValue.rawValue, forKey: .consoleSwitches)
+// MARK: Custom functionality
+private extension RayRacerDelegate {
+	func withModalFileOpenPanel(_ perform: (URL) -> Void) {
+		let panel = NSOpenPanel()
+		panel.allowsMultipleSelection = false
+		panel.canChooseFiles = true
+		panel.canChooseDirectories = false
+		panel.canCreateDirectories = false
+		panel.directoryURL = self.defaults.openedFileURLs.first
+		
+		let response = panel.runModal()
+		if let url = panel.url,
+		   response == .OK {
+			perform(url)
 		}
 	}
 	
-	var openedFileURLs: [URL] {
-		// show up to 10 recently opened files
-		return self.openedFileBookmarks
-			.prefix(10)
-			.map({ $0.0 })
-	}
-	
-	func addOpenedFileURL(_ url: URL) {
-		guard let data = try? url.bookmarkData(options: .readOnlySecurityScope) else {
-			return
+	func openFile(at url: URL) {
+		guard let data = try? Data(contentsOfSecurityScopedResourceAt: url) else {
+			// TODO: show error when opening cartridge data fails
+			fatalError()
 		}
 		
-		// read bookmark data in user defaults, excluding bookmark data of
-		// the new URL; read bookmark data of 9 recently opened files to
-		// limit the result to 10, once the new data is added
-		var defaultsData = self.openedFileBookmarks
-			.filter({ $0.0 != url })
-			.prefix(9)
-			.map({ $0.1 })
+		let viewController = ScreenViewController(
+			commandQueue: self.commandQueue,
+			pipelineState: self.pipelineState)
 		
-		// prepend bookmark data of the new URL at the beginning and write
-		// bookmark data to user defaults
-		defaultsData.insert(data, at: 0)
-		self.setValue(defaultsData, forKey: .openedFileBookmarks)
-	}
-	
-	func clearOpenedFileURLs() {
-		self.removeObject(forKey: .openedFileBookmarks)
-	}
-	
-	private var openedFileBookmarks: [(URL, Data)] {
-		guard let data = self.value(forKey: .openedFileBookmarks) as? [Data] else {
-			return []
-		}
+		let windowController = NSWindowController(windowNibName: "ScreenWindow")
+		windowController.contentViewController = viewController
+		windowController.window?
+			.title = url.lastPathComponent
 		
-		// resolve file URLs from bookmark data and only keep unique ones
-		var bookmarks: [(URL, Data)] = []
-		for data in data {
-			var stale = false
-			
-			// resolve file URLs from bookmark data and only keep unique ones
-			if let url = try? URL(resolvingBookmarkData: data, options: .securityScope, relativeTo: nil, bookmarkDataIsStale: &stale),
-			   bookmarks.contains(where: { $0.0 == url }) == false {
-				bookmarks.append((url, data))
-			}
-			
-			if stale {
-				// TODO: update opened file URL stale bookmark
-			}
-		}
+		self.console.tia.output = viewController
+		self.console.switches = self.defaults.consoleSwitches
+		self.console.insertCartridge(data)
+		self.console.reset()
 		
-		return bookmarks
+		self.showWindow(of: windowController)
+		self.defaults.addOpenedFileURL(url)
 	}
-}
-
-private extension URL.BookmarkCreationOptions {
-	static let readOnlySecurityScope: Self = [
-		.withSecurityScope,
-		.securityScopeAllowOnlyReadAccess
-	]
-}
-
-private extension URL.BookmarkResolutionOptions {
-	static let securityScope: Self = [
-		.withSecurityScope,
-		.withoutUI
-	]
 }
